+++ car/car.h
@@ -2,6 +2,6 @@
 #define Car_h
 #include <stdint.h>
 
-void setCarMovement(int32_t leftWheels, int32_t rightWheels);
+void setCarMovement(int8_t leftWheels, int8_t rightWheels);
 void setupCar();
 #endif
